♻️ Single instance
